# Enhance ArchGDAL OGR objects with parametric composite types