Here is a draft for review, demonstrating:
1. a way to dispatch on type in `getfield` (to fix issue #246) bring for a 20% performance gain on layer to table conversion
2. the use of `ogr_f_stealgeometry` for a gain of an additionnal 20%

<details><summary>Performance results</summary><p>
```julia
julia> using ArchGDAL
       const AG = ArchGDAL
       using BenchmarkTools
       using Tables
       using Shapefile

julia> ds = AG.read("../../tmp/road/road.shp")
       layer = AG.getlayer(ds, 0)
       fdp_layer = AG.getFDPlayer(ds, 0)
Layer: road
  Geometry 0 (): [wkbLineString], LINESTRING (874644.4...), ...
     Field 0 (gid): [OFTInteger], 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, ...
     Field 1 (roadcode): [OFTString], 01A903905CD, 01A903905CD, 01A903905CD, ...
     Field 2 (plod): [OFTString], DB1, DB2, DB3, DB4, DB1, DB2, DB1, DB2, ...
     Field 3 (absd): [OFTReal], 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, ...
     Field 4 (plof): [OFTString], FB1, FB2, FB3, FB4, FB1, FB2, FB1, FB2, ...
...
 Number of Fields: 7

julia> @benchmark Tables.columns(Shapefile.Table("../../tmp/road/road.shp")
       )
BenchmarkTools.Trial: 71 samples with 1 evaluation.
 Range (min … max):  65.360 ms … 79.845 ms  ┊ GC (min … max): 0.00% … 12.70%
 Time  (median):     68.829 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   70.556 ms ±  3.834 ms  ┊ GC (mean ± σ):  4.00% ±  4.75%

    ▂ ▂  ▂ █  █  █▂                    ▂   █  ▂                
  ▅▁█▅██▅███▁███▅██▅▁▅▁▁▅▁▁▅▁▁▁▁▅▅▁▁▅▁▅█▁▅██▁▅██▅▁▁▁▁▁█▅▁▁▁▁█ ▁
  65.4 ms         Histogram: frequency by time        78.2 ms <

 Memory estimate: 25.48 MiB, allocs estimate: 54329.

julia> @benchmark Tables.columns($layer)
BenchmarkTools.Trial: 42 samples with 1 evaluation.
 Range (min … max):   97.639 ms … 230.930 ms  ┊ GC (min … max): 0.00% … 20.99%
 Time  (median):     101.620 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   120.142 ms ±  42.069 ms  ┊ GC (mean ± σ):  5.93% ±  7.61%

  ▂█▁                                                            
  ███▄▁▃▁▁▁▁▁▁▃▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▃▁▁▁▁▃▁▁▁▃▁▁▄▁▁▁▁▁▃▃ ▁
  97.6 ms          Histogram: frequency by time          231 ms <

 Memory estimate: 8.98 MiB, allocs estimate: 373870.

julia> @benchmark Tables.columns($fdp_layer)
BenchmarkTools.Trial: 53 samples with 1 evaluation.
 Range (min … max):  81.740 ms … 163.887 ms  ┊ GC (min … max): 0.00% … 9.56%
 Time  (median):     87.945 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   97.217 ms ±  24.621 ms  ┊ GC (mean ± σ):  2.62% ± 3.99%

    ▄▇█▇                                                        
  ▅█████▅▅▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▃▃▁▃▁▁▁▁▅▁▁▁▅ ▁
  81.7 ms         Histogram: frequency by time          161 ms <

 Memory estimate: 7.75 MiB, allocs estimate: 279326.

```
</p></details>

## 1. ArchGDAL OGR related types moved to parametric composite types

Trying to dispatch on types required: 

### 1.1 Creation of two parametric singleton types for fields and geomfields, and of a featuredefn like type alias
- **`FType`** for fields, which is paramaterized by a tuple of `OGRFieldType` and `OGRFieldSubType`
- **`GType`** for geomfields, which is parameterized by an `OGRwkbGeometryType`
- **`FDType`** for featuredefn, which is a 2-tuple of a tuple of `GType` and a tuple of `FType`

_Note 1: I have added convert functions from `FType` and `DataType` which yields a more readable correspondance between Julia DataTypes and GDAL Field type ids, than the current separated convert functions from `OGRFieldType` and `OGRFieldSubType` to `DataType`_

_Note 2: I have also added a commented draft of new `Geometry` parametric singleon type (named `Geom`) based on `GType` but not used it since I have not identified any benefit_

### 1.2 Duplication of most of OGR related composite types and parameterization with `FType`, `GType` and `FDType`
The duplicated types have the `FDP` prefix
- **`FDP_GeomFieldDefn`** and **`FDP_FieldDefn`** types are parameterized by `GType` and `FType`
- **`FDP_FeatureDefn`**, **`FDP_Feature`** and **`FDP_FeatureLayer`** types are parameterized by `FDType`  

All of them have been set as subtypes of corresponding abstract supertypes **`DUAL_AbstractXXX`**, in order to limit the duplication of functions code when not necessary.

_Note: A plain FeatureDefn parametric composite type could later be used as type parameter for Feature and FeatureLayer_ 

### 1.3 Modification of a few functions, used to show a layer and convert it to a table
- `getFDPlayer` to handle a feature layer data source with all the newly introduced parametric composite types
- Some functions, which do not have any gain on using the `FDType` information, have just been **extended to accept `DUAL_AbstractXXX`**
- Most of the functions benefiting for the `FDType` information, have been **duplicated as generated functions**

_Note: In the event of the **unlikely use case where there would be too many generated functions** (a lot of different data sources with different featuredefn) for compilation to be performed in a reasonable time or even to succeed, the **generated functions could be moved to optionally-generated functions**. The fallbacks being the legacy functions_

### 1.4 Side additions
1. I have added a macro to export `Enum` type id for user convenience. Disabled in the PR because it is not convenient when developping using revise.jl because it redefines the related constants. I have not found a way to inhibit the latter
2. I have generalized the `ownedby` field in feature layer related types which may serve as a basis to fix issue #215, but @yeesian is tackling this one. For all intents and purposes here are below the relationships which are proposed:  
   <details><summary>Ownership relations graph</summary><p>
   ![OGR Julia objects ownership relations diagram V1]()
   </p></details>

   `destroy` functions for `FDP_XXX` types have been created accordingly


## 2. Steal geometry instead of cloning for the first geometry in features
There appears that `Tables.rows(layer)` makes a copy of the layer.  
Since `ogr_f_stealgeometry` is faster than the composition of `ogr_f_getgeomfieldref` and `ogr_g_clone`, I have modified `Tables.getcolumn` to use a `stealgeom` function instead of the current `getgeom` function.

Unfortunately, `ogr_f_getgeomfieldref` is implemented in GDAL C interface only for the default (first) geometry. Datasources with only one geometry per feature is the general case, but that could be an issue to raise in GDAL. 

## 3. Further performance optimizations to investigate, regarding the layer to table conversion performance
- Include field and geomfieldnames in feature definition type `FDType` to avoid calling `ogr_f_getfieldindex`
- Count the number of calls to `ogr_l_getnextfeature` to ensure there is no unnecessary call to it