@mathieu17g 2021-11-26
> - Include field and geomfield names in feature definition type `FDType` to avoid calling `ogr_f_getfieldindex`

I have changed the FeatureDefn type parameter of ArchGDAL OGR objects from 

```julia
FDType = Tuple{NTuple{NG,GType} where NG,NTuple{NF,FType} where NF}
```
which is the same as
```julia
FDType = Tuple{
    Tuple{Vararg{GType, NG}} where NG, 
    Tuple{Vararg{FType, NF}} where NF,
}
```
to (thanks to https://discourse.julialang.org/t/parametric-type-with-a-constrained-namedtuple/34511)
```julia
FDType = Tuple{
    NamedTuple{NG,<:Tuple{Vararg{GType}}} where NG,
    NamedTuple{NF,<:Tuple{Vararg{FType}}} where NF,
}
```

It brings some performance gain by avoiding most of the calls to `ogr_f_getfieldindex`:
Current master:
```julia
@benchmark Tables.columntable(AG.getlayer(AG.read("../../tmp/road/road.shp"), 0)) seconds = 30
BenchmarkTools.Trial: 142 samples with 1 evaluation.
 Range (min … max):  182.404 ms … 335.852 ms  ┊ GC (min … max): 0.00% … 10.29%
 Time  (median):     196.571 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   213.022 ms ±  37.221 ms  ┊ GC (mean ± σ):  3.27% ±  4.88%

   ▃█▅▄▂▄  ▄                                                     
  ▆██████▇██▇▃▃▁▃▁▃▁▃▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▃▁▃▁▁▃▃▅▄▇▁▃▄▆▃▅▃▁▁▁▁▁▁▁▁▁▃ ▃
  182 ms           Histogram: frequency by time          318 ms <

 Memory estimate: 8.98 MiB, allocs estimate: 373869.
```
Shapefile.jl:
```julia
julia> @benchmark Tables.columntable(Shapefile.Table("../../tmp/road/road.shp")) seconds = 30BenchmarkTools.Trial: 198 samples with 1 evaluation.
 Range (min … max):  142.824 ms … 163.894 ms  ┊ GC (min … max): 0.00% … 5.48%
 Time  (median):     152.970 ms               ┊ GC (median):    4.10%
 Time  (mean ± σ):   152.064 ms ±   4.070 ms  ┊ GC (mean ± σ):  2.95% ± 2.46%

                  ▃▂▂▄ ▄               ▇▂█▂▄▄▅                   
  ▃▁▁▃▁▃▃▅▆▅▆▆▅█▇█████▅██▃▃▅▃▃▅▁█▆██▆▇████████▆▃▆▇▇▆▅▅▁▁▁▃▃▁▁▁▃ ▃
  143 ms           Histogram: frequency by time          161 ms <

 Memory estimate: 35.89 MiB, allocs estimate: 243040.
```
Current master with geometry stealing:
```julia
julia> @benchmark Tables.columntable(AG.getlayer(AG.read("../../tmp/road/road.shp"), 0)) seconds = 30
BenchmarkTools.Trial: 144 samples with 1 evaluation.
 Range (min … max):  178.243 ms … 312.567 ms  ┊ GC (min … max): 0.00% … 13.64%
 Time  (median):     191.755 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   208.443 ms ±  40.801 ms  ┊ GC (mean ± σ):  3.39% ±  5.23%

    █▂▁▆▅▂▂                                                      
  ▃████████▅▄▃▃▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▃▁▁▃▁▁▃▃▃▄▃▅▁▃▄▃▃▃▃ ▃
  178 ms           Histogram: frequency by time          312 ms <

 Memory estimate: 8.98 MiB, allocs estimate: 373879.
```
New ArchGDAL OGR types with geometry stealing:
```julia
julia> @benchmark Tables.columntable(AG.getFDPlayer(AG.read("../../tmp/road/road.shp"), 0)) seconds = 30
BenchmarkTools.Trial: 169 samples with 1 evaluation.
 Range (min … max):  158.547 ms … 245.862 ms  ┊ GC (min … max): 0.00% … 7.86%
 Time  (median):     169.471 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   178.152 ms ±  22.915 ms  ┊ GC (mean ± σ):  1.48% ± 2.67%

   ▁▃▃▆▄▇█▁▃▂▆▁                                                  
  ▇████████████▇▄▃▁▁▃▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▃▁▁▁▁▁▃▅▇▃▃▆▄▅▃▃▁▃▁▁▁▃ ▃
  159 ms           Histogram: frequency by time          243 ms <

 Memory estimate: 7.75 MiB, allocs estimate: 279394.
```
Beyond the call to `ogr_l_getnextfeature`, the last step taking some time is the `IGeometry` constructor: < 10% of the total time