@mathieu17g 2021-11-26
@yeesian I'm not able to push the commit associated to my previous comment into branch Enhance_ArchGDAL_types. I got the following message:
```
remote: error: GH006: Protected branch update failed for refs/heads/Enhance_ArchGDAL_types.
remote: error: At least 1 approving review is required by reviewers with write access.
To https://github.com/yeesian/ArchGDAL.jl.git
! [remote rejected] Enhance_ArchGDAL_types -> Enhance_ArchGDAL_types (protected branch hook declined)
error: failed to push some refs to 'https://github.com/yeesian/ArchGDAL.jl.git'
```

Do you know what I did wrong ?