@yeesian 2021-11-26
I've updated the branch protection settings now, can you try again? Thanks!