@mathieu17g 2021-11-26
It stills fails, but I have a different message:
```
remote: error: GH006: Protected branch update failed for refs/heads/Enhance_ArchGDAL_types.
remote: error: Changes must be made through a pull request.
To https://github.com/yeesian/ArchGDAL.jl.git
! [remote rejected] Enhance_ArchGDAL_types -> Enhance_ArchGDAL_types (protected branch hook declined)
error: failed to push some refs to 'https://github.com/yeesian/ArchGDAL.jl.git'
```