@mathieu17g 2021-12-01
Continuing to shave off time in layer to table conversion, I noticed that `_infergeomtype` could be optimized by modifying convert functions between GDAL CEnum.Cenum types to AG Enum types.

### New convert functions
```julia
function convert(::Type{OGRwkbGeometryType}, gogtinst::GDAL.OGRwkbGeometryType)
    return OGRwkbGeometryType(Integer(gogtinst))
end
function convert(::Type{GDAL.OGRwkbGeometryType}, ogtinst::OGRwkbGeometryType)
    return GDAL.OGRwkbGeometryType(Integer(ogtinst))
end
```
The convert functions code looks simpler and faster like this, and could be extended to other similar conversions between ArchGDAL Enum and GDAL CEenum.Cenum

This has the prerequisite to align `OGRwkbGeometryType` Enum instances' assigned values to those of `GDAL.OGRwkbGeometryType`. It should not have any side effect.

### Results
- With current master:
```julia
julia> geom = AG.fromWKT("POINT (1 2)")
Geometry: POINT (1 2)

julia> @benchmark AG._infergeomtype($(geom.ptr))
BenchmarkTools.Trial: 10000 samples with 949 evaluations.
 Range (min … max):  83.511 ns … 209.497 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     97.937 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   95.838 ns ±  13.089 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  █ ▆ ▂▂       ▄█▂▁▅ ▂▂  ▁▁▁▃▁▁          ▁     ▂               ▂
  █▂█▄██▆▆▅▄█▆██████████▇███████▇▇▇█▇▇▇▇▇██▆▆▇▇█▆▆▆▇▆▅▆▆▄▅▅▆▅▆ █
  83.5 ns       Histogram: log(frequency) by time       143 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.

```
- With new convert functions:
```julia
julia> geom = AG.fromWKT("POINT (1 2)")
Geometry: POINT (1 2)

julia> @benchmark AG._infergeomtype($(geom.ptr))
BenchmarkTools.Trial: 10000 samples with 997 evaluations.
 Range (min … max):  18.790 ns … 78.432 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     19.063 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   19.519 ns ±  2.273 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▇█▄ ▅  ▁                      ▂ ▂               ▁           ▂
  ███▅██▄█▃▅▄▃▄▁▁▄▃▁▃▁▃▁▁▁▁▁▃▁▁▁█▆█▁▁▁▁▁▁▁▁▁▃▁▁▆█▅█▆▆▇▆▅▅▆▆▆▇ █
  18.8 ns      Histogram: log(frequency) by time      29.9 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.

```
### Next
I will go saving time with a trial to implement a specialized version of `Tables.eachcolumns` to strip some time from calls to `ArchGDAL.getfield`